app/requirements.txt
Flask>=3.0.0
Flask-Session>=0.8.0
msal>=1.35.0
openai>=2.21.0
python-dotenv>=1.0.1
requests>=2.32.0
Werkzeug>=3.1.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0
gunicorn>=21.0.0